+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Fixes
 
+- Don't send logs and metrics when the SDK is disabled (#8173)
 - Fixes crash caused by modifying breadcrumbs from multiple threads (#8114)
 - Prevent feedback form on external displays (#8071)
 - Keep the User Feedback screenshot trigger active after form dismissal. (#8048)